refactor: reduced type complexity on UI components
